+++ src/api/time.rs
@@ -175,6 +175,9 @@ mod tests {
     use std::thread;
     use std::time::Duration;
 
+    // TODO
+    // Use MOCK Clock for test timings because they are manipulated by CPU load or VMs etc
+
     #[test]
     fn new_stopwatch_initial_state() {
         let sw = Stopwatch::new();
@@ -189,7 +192,8 @@ mod tests {
         sw.start();
         thread::sleep(Duration::from_millis(100));
         let elapsed = sw.read();
-        assert!(elapsed >= 0.09 && elapsed <= 0.2, "elapsed = {}", elapsed);
+        // assert!(elapsed >= 0.09 && elapsed <= 0.2, "elapsed = {}", elapsed);
+        assert!(elapsed >= 0.09, "elapsed = {}", elapsed);
     }
 
     #[test]