+++ src/helper/mod.rs
@@ -12,22 +12,22 @@ use crate::helper::print::{BLUE, BOLD, BRIGHT_BLUE, CYAN, END, GREEN, PURPLE, RE
 // add no color argument for no color output
 
 //[GENERALL-OPTIONS]
-const sco: &str = PURPLE;
+const SCO: &str = PURPLE;
 
 //[SCRIPTOPTIONS]
 const SC: &str = YELLOW;
 
 //[OPTIONS]
-const opt: &str = CYAN;
+const OPT: &str = CYAN;
 
 //<action>
 const AC: &str = RED;
 
 //<custom-action-options>
-const aco: &str = CYAN;
+const ACO: &str = CYAN;
 
 // OptionList
-const op: &str = BLUE;
+const OP: &str = BLUE;
 
 // LINKs
 const LINK: &str = BLUE;
@@ -43,61 +43,61 @@ pub fn help() {
     // Lua Scripts
     println!(
         "\nUsage for Lua scripts: {}<flua>{} <script.lua> {}[SCRIPTOPTIONS]{} {}[GENERALL-OPTIONS]{}",
-        GREEN, END, SC, END, sco, END
+        GREEN, END, SC, END, SCO, END
     );
     //[SCRIPTOPTIONS]
     println!("\n{}[SCRIPTOPTIONS]{}", SC, END);
     println!(
         "{}  --safe:        {}Run in safe mode (limited API, no OS access)",
-        op, END
+        OP, END
     );
     println!(
         "{}  --no-info:     {}Suppress start and end info messages",
-        op, END
+        OP, END
     );
     println!(
         "{}  l, lua-args    {}submit arguments after lua-args for the lua file which will be run, stop the collectiong when it sees an argument which starts with: {}'-'{}",
-        op, END, RED, END
+        OP, END, RED, END
     );
     // Modules
     println!(
         "\nUsage for Modules: {}<flua>{} run {}<actions>{} {}<custom-action-options>{} {}[GENERALL-OPTIONS]{}",
-        GREEN, END, AC, END, aco, END, sco, END
+        GREEN, END, AC, END, ACO, END, SCO, END
     );
     println!("\n{}<actions>{}:   'update', 'install'", AC, END);
     println!(
         "{}  module         {}Argument to run a {}dlm13{} Module",
-        op, END, RED, END
+        OP, END, RED, END
     );
     println!(
         "\n{}<custom-action-options>{}: Ends the arg collecting when a path starts with #",
-        aco, END
+        ACO, END
     );
-    println!("{}  -path=    {}Add the Module path after the '='", op, END);
+    println!("{}  -path=    {}Add the Module path after the '='", OP, END);
     // Other
     println!(
         "\nOther Usage: {}<flua>{} {}[OPTIONS]{} {}[GENERALL-OPTIONS]{}",
-        GREEN, END, opt, END, sco, END
+        GREEN, END, OPT, END, SCO, END
     );
     //[OPTIONS]
-    println!("\n{}[OPTIONS]{}", opt, END);
+    println!("\n{}[OPTIONS]{}", OPT, END);
     println!(
         "{}  -v, --version  {}Function which prints the version in the terminal",
-        op, END
+        OP, END
     );
     println!(
         "{}  h, -h, --help  {}Function which prints this help message in the terminal",
-        op, END
+        OP, END
     );
     //[GENERALL-OPTIONS]
-    println!("\n{}[GENERALL-OPTIONS]{}", sco, END);
+    println!("\n{}[GENERALL-OPTIONS]{}", SCO, END);
     println!(
         "{}  -nw            {}No exit -> The Programm closes instantly after an error and showing an error message without waiting for some Time.",
-        op, END
+        OP, END
     );
     println!(
         "{}  -no-config     {}The Programm will not search or load the Config file. Standard Option is true, if there is no Config file, the step will be skipped. A Config file will not be created automaticly!",
-        op, END
+        OP, END
     );
     // CONFIG
     println!("\n{}[MORE-OPTIONS]{}", CONFIG, END);
@@ -114,32 +114,32 @@ pub fn config_help() {
     println!("{}[CONFIG]{} Help Info for Flua v{}", CONFIG, END, VERSION);
     println!(
         "\nUsage: flua conifg {}[CONFIGOPTIONS]{} {}[GENERALL-OPTIONS]{}",
-        CONFIG, END, sco, END
+        CONFIG, END, SCO, END
     );
     println!("\n{}[CONFIGOPTIONS]{}", CONFIG, END);
     println!(
         "{}    generate{}    Be careful! This will create a new config file, the old one will be overritten",
-        op, END
+        OP, END
     );
     println!(
         "{}    open{}        to open the config file in the default programm",
-        op, END
+        OP, END
     );
     println!(
         "{}    opendir{}     to open the config file in the default programm",
-        op, END
+        OP, END
     );
     println!(
         "{}    dir{}         to open the config file in the default programm",
-        op, END
+        OP, END
     );
     println!(
         "{}    check{}       to check if the config file is correct",
-        op, END
+        OP, END
     );
     println!(
         "{}    clean{}       to delete the current configfile",
-        op, END
+        OP, END
     );
 }