+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - loading the Config File
 - added config file Infos to the help message
 - added Tests for Time
+- added Config Options
+- added Input functions like `input()` in Python
 
 ## 0.2.0